+++ src/models/blog.model.js
@@ -12,11 +12,11 @@ module.exports = (mongoose) => {
     { timestamps: true }
   );
 
-  // schema.method("toJSON", function () {
-  //   const { __v, _id, ...object } = this.toObject();
-  //   object.id = _id;
-  //   return object;
-  // });
+  schema.method("toJSON", function () {
+    const { __v, _id, ...object } = this.toObject();
+    object.id = _id;
+    return object;
+  });
 
   const blog = new mongoose.model("blog", schema);
   return blog;
